Message of ConnecterDataBase

+++ tourists/helpers/TrainsHelper.java
@@ -3,6 +3,8 @@ package tourists.helpers;
 import java.util.*;
 import java.io.*;
 
+import tourists.StringMaster;
+
 public class TrainsHelper implements QueryHelper{
 	@Override
 	public String getSelectingQuery(Map<String, String> fields, List<String> flags){
@@ -29,18 +31,44 @@ public class TrainsHelper implements QueryHelper{
 			return null;
 		}
 		StringBuilder query = new StringBuilder("INSERT INTO TRAINS VALUES(");
-		if(values.containsKey("COACH_NAME") && values.containsKey("COACH_LAST_NAME") && values.containsKey("COACH_BIRTH")){
-			query.append("(SELECT ID FROM TOURISTS WHERE NAME='" + values.get("COACH_NAME") + "' AND ");
-			query.append("LAST_NAME='" + values.get("COACH_LAST_NAME") + "' AND ");
-			query.append("BIRTH='" + values.get("COACH_BIRTH") + "')");
+		if(values.containsKey("COACH_NAME") || values.containsKey("COACH_LAST_NAME")
+			|| values.containsKey("COACH_BIRTH")){
+			query.append("(SELECT ID FROM TOURISTS WHERE ");
+			String name = values.get("COACH_NAME");
+			String lastName = values.get("COACH_LAST_NAME");
+			String birth = values.get("COACH_BIRTH");
+			if(!StringMaster.isDate(birth)){
+				System.err.println(birth + " is not a date");
+				return null;
+			}
+			if(name != null){
+				query.append("NAME='" + name + "' AND ");
+			}
+			if(lastName != null){
+				query.append("LAST_NAME='" + lastName + "' AND ");
+			}
+			if(birth != null){
+				query.append("BIRTH='" + birth.substring(0, DATE_LENGTH) + "' AND ");
+			}
+			query.delete(query.length() - " AND ".length(), query.length());
+			query.append(")");
+		}
+		else{
+			query.append("NULL");
 		}
 		query.append(",");
 		if(values.containsKey("GROUP_ID")){
 			query.append(values.get("GROUP_ID"));
 		}
+		else{
+			query.append("NULL");
+		}
 		query.append(",");
 		if(values.containsKey("TRAINING")){
 			query.append("(SELECT ID FROM TRAININGS WHERE NAME='" + values.get("TRAINING") + "')");
+		}
+		else{
+			query.append("NULL");
 		}
 		query.append(")");
 		return query.toString();
@@ -52,32 +80,46 @@ public class TrainsHelper implements QueryHelper{
 			return null;
 		}
 		StringBuilder query = new StringBuilder("UPDATE TRAINS SET ");
-		values.forEach((String attribute, String value)->{
-			switch(attribute){
-				case "GROUP_ID":
-					query.append(attribute + "=" + value + ",");
-					break;
-				case "TRAINING":
-					query.append(attribute + "=(SELECT ID FROM TRAININGS WHERE NAME='" + value + "'),");
-					break;
+		query.append("GROUP_ID=");
+		if(values.containsKey("GROUP_ID")){
+			query.append(values.get("GROUP_ID") + ",");
+		}
+		else{
+			query.append("NULL,");
+		}
+		query.append("TRAINING=");
+		if(values.containsKey("TRAINING")){
+			query.append("(SELECT ID FROM TRAININGS WHERE NAME='" + values.get("TRAINING") + "'),");
+		}
+		else{
+			query.append("NULL,");
+		}
+		query.append("COACH=");
+		if(values.containsKey("COACH_NAME") || values.containsKey("COACH_LAST_NAME")
+			|| values.containsKey("COACH_BIRTH")){
+			query.append("(SELECT ID FROM TOURISTS WHERE ");
+			String name = values.get("COACH_NAME");
+			String lastName = values.get("COACH_LAST_NAME");
+			String birth = values.get("COACH_BIRTH");
+			if(!StringMaster.isDate(birth)){
+				System.err.println(birth + " is not a date");
+				return null;
 			}
-		});
-		query.append("COACH=(SELECT ID FROM TOURISTS WHERE ");
-		values.forEach((String attribute, String value)->{
-			switch(attribute){
-				case "COACH_NAME":
-					query.append("NAME='" + value + "' AND ");
-					break;
-				case "COACH_LAST_NAME":
-					query.append("LAST_NAME='" + value + "' AND ");
-					break;
-				case "COACH_BIRTH":
-					query.append("BIRTH='" + value + "' AND ");
-					break;
+			if(name != null){
+				query.append("NAME='" + name + "' AND ");
 			}
-		});
-		query.delete(query.length() - " AND ".length(), query.length());
-		query.append(")");
+			if(lastName != null){
+				query.append("LAST_NAME='" + lastName + "' AND ");
+			}
+			if(birth != null){
+				query.append("BIRTH='" + birth.substring(0, DATE_LENGTH) + "' AND ");
+			}
+			query.delete(query.length() - " AND ".length(), query.length());
+			query.append(")");
+		}
+		else{
+			query.append("NULL");
+		}
 		query.append("\nWHERE ");
 		fields.forEach((String attribute, String value)->{
 			switch(attribute){
@@ -166,5 +208,6 @@ public class TrainsHelper implements QueryHelper{
 		return text.toString();
 	}
 	
+	private int DATE_LENGTH = 10;
 	private String SELECT_FILE = "SQL_select_trains.txt";
 }
